Fix clang compilation warning on macOS

+++ src/fileio.cpp
@@ -146,7 +146,7 @@ char* dvz_parse_npy(DvzSize size, char* npy_bytes)
     }
 
     // Extract the header length (at byte 8 and 9 for v1.0/1.1)
-    uint16_t header_len = *(uint16_t*)(npy_bytes + 8);
+    uint16_t header_len = *(uint16_t*)((uint64_t)npy_bytes + 8);
 
     // Calculate the offset of the array data
     DvzSize data_offset = 10 + header_len;
